feat: add endpoint to get entrypoint params used in swappable graph
To support entrypoint task swaps, we need to a way to retrieve a dynamic list of global params that
is dependent on which tasks are specified for each swap. This will allow us to notify the user of
any unused global parameters.

The implementation parses an entrypoint graph YAML and cross reference declared global params for
the entrypoint with the dereferenced variable names in the graph YAML.

The new endpoint is at `/entrypoints/{id}/snapshots/{snapshotId}/dynamicGlobalParameters`. It
returns the list of entrypoint params for the snapshot with the provided swaps, the topologically
sorted graph steps, and the list of active plugins.

The Python client has been updated for the new endpoint.

This commit includes tests for the new functionality.